refactor(auto-continue): widen e2e scheduled window + share test helpers
- Widen scheduledAt offset to 10s to avoid real-clock flakiness on slow runners
- Drop redundant setTimeout(0) drains — waitFor already handles async
- Extract AsyncEventQueue + waitFor to src/server/test-helpers/ (shared by agent.test and e2e.test)
- Add optional label arg to waitFor for clearer timeout messages
- Shut down ScheduleManager in test finally
